Move the composer and component into same folder structure as other projects
