+++ src/SeoToolkit.Umbraco.Deploy/Components/DeployComposer.cs
@@ -1,0 +1,13 @@
+﻿using Umbraco.Cms.Core.Composing;
+using Umbraco.Cms.Core.DependencyInjection;
+
+namespace SeoToolkit.Umbraco.Deploy.Components
+{
+    public class DeployComposer : IComposer
+    {
+        public void Compose(IUmbracoBuilder builder)
+        {
+            builder.AddComponent<DeployComponent>();
+        }
+    }
+}
